fix: adjust iframe background in dark mode
in dark mode the text of ttl / markdown documents appeared white on white in some browsers rendering it effectively invisible. This makes sure the background is also adjusted. It seems not to change anything for browsers that do not style the iframe colors anyway. So it is either light on dark or dark on light in any browser.